feat(arch): extract agent integrations from server/
Move codex handlers/validators and OpenCode hooks out of src/server/
into a new src/integrations/ tree. This is the first of 6 commits in
the architecture refactor (see docs/REFACTOR-2026-04-architecture.md).

- codex/* moves to src/integrations/codex/
- hooks/* moves to src/integrations/opencode/hooks/

No port definition yet (Commit 3); codex routes still listed in the
central routes.ts table (transient until Commit 3 moves them via the
AgentIntegration port).

Tests stay co-located with sources per AGENTS.md §3.
