+++ src/integrations/opencode/hooks/index.ts
